M3-1435: Clean code
